The streak page is categories now: history, streak table, freezes
Two tiles, a share button, sixty days of history and a two-tab board under
it that nobody scrolled to. Same shape as the wallet now: the numbers and
the share at the top, then a row per thing to look at, each its own page.
The freezes row carries the banked count and opens the store, because "how
do I protect this" is asked here and answered there.
